Updated tuners to work on GPU backends

+++ src/MorpheusOracle_DecisionTreeTuner.hpp
@@ -129,9 +129,9 @@ class DecisionTreeTuner {
    * that until the tuner finishes the tuning process, the format ID is set to
    * \p INVALID_FORMAT_STATE.
    *
-   * @return size_t Optimum format index.
+   * @return int Optimum format index.
    */
-  size_t format_id() const { return format_id_; }
+  int format_id() const { return format_id_; }
 
   /**
    * @brief Checks if the tuner prints verbose messages.
@@ -152,7 +152,7 @@ class DecisionTreeTuner {
  private:
   DecisionTree tree_;
   scalar_vector timings_;
-  size_t format_id_;
+  int format_id_;
   bool verbose_;
   /*! \endcond */
 };
